double i

+++ src/provagithub/ProvaGitHub.java
@@ -15,7 +15,7 @@ public class ProvaGitHub {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        int i=0;
+        double i=0;
     while(true){
         System.out.println("ciao: "+i); 
         i++;
